requirements.txt
python-dotenv==1.0.1
pandas==2.2.1
requests==2.31.0
beautifulsoup4==4.12.3
